Spin out `oak_package_metadata`
To avoid circular dependencies between oak_package and oak_sources in upcoming changes
