+++ crates/ark/Cargo.toml
@@ -82,6 +82,7 @@ ark_test.workspace = true
 assert_matches.workspace = true
 insta.workspace = true
 oak_package_definitions = { workspace = true, features = ["testing"] }
+oak_package_metadata.workspace = true
 stdext = { workspace = true, features = ["testing"] }
 tempfile.workspace = true
