+++ crates/oak_ide/Cargo.toml
@@ -31,4 +31,5 @@ assert_matches.workspace = true
 oak_index.workspace = true
 oak_package = { workspace = true, features = ["testing"] }
 oak_package_definitions = { workspace = true, features = ["testing"] }
+oak_package_metadata.workspace = true
 oak_sources = { workspace = true, features = ["testing"] }
