+++ crates/oak_ide/tests/goto_definition.rs
@@ -16,9 +16,9 @@ use oak_layers::scope_layer::file_layers;
 use oak_layers::scope_layer::ScopeLayer;
 use oak_package::library::Library;
 use oak_package::package::Package;
-use oak_package::package_description::Description;
-use oak_package::package_namespace::Namespace;
 use oak_package_definitions::LibraryDefinitions;
+use oak_package_metadata::description::Description;
+use oak_package_metadata::namespace::Namespace;
 use oak_sources::test::TestPackageCache;
 use stdext::SortedVec;
 use url::Url;
