+++ crates/oak_layers/tests/external.rs
@@ -13,10 +13,10 @@ use oak_layers::scope_layer::package_root_layers;
 use oak_layers::scope_layer::ScopeLayer;
 use oak_package::library::Library;
 use oak_package::package::Package;
-use oak_package::package_description::Description;
-use oak_package::package_namespace::Import;
-use oak_package::package_namespace::Namespace;
 use oak_package_definitions::LibraryDefinitions;
+use oak_package_metadata::description::Description;
+use oak_package_metadata::namespace::Import;
+use oak_package_metadata::namespace::Namespace;
 use oak_sources::test::TestPackageCache;
 use stdext::SortedVec;
 use url::Url;
