+++ crates/oak_package/src/lib.rs
@@ -1,7 +1,2 @@
 pub mod library;
 pub mod package;
-pub mod package_description;
-pub mod package_index;
-pub mod package_namespace;
-
-pub use package_description::Dcf;
