+++ crates/oak_package_definitions/src/definitions.rs
@@ -7,7 +7,7 @@ use biome_text_size::TextRange;
 use oak_index::semantic_index;
 use oak_index_vec::define_index;
 use oak_index_vec::IndexVec;
-use oak_package::package_namespace::Namespace;
+use oak_package_metadata::namespace::Namespace;
 use rustc_hash::FxHashMap;
 use stdext::result::ResultExt;
