+++ crates/oak_package_metadata/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "oak_package"
+name = "oak_package_metadata"
 version = "0.1.0"
 description = """
-R package metadata: DESCRIPTION, NAMESPACE, INDEX readers and Library cache.
+R package metadata: DESCRIPTION, NAMESPACE, INDEX readers.
 """
 
 authors.workspace = true
@@ -17,9 +17,12 @@ testing = ["dep:tempfile"]
 workspace = true
 
 [dependencies]
+aether_parser.workspace = true
+aether_syntax.workspace = true
+biome_rowan.workspace = true
 anyhow.workspace = true
-log.workspace = true
-oak_package_metadata.workspace = true
+oak_core.workspace = true
+regex.workspace = true
 stdext.workspace = true
 
 [dependencies.tempfile]
