+++ crates/oak_package_metadata/src/index.rs
