+++ crates/oak_package_metadata/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod dcf;
+pub mod description;
+pub mod index;
+pub mod namespace;
