+++ crates/oak_package_metadata/src/namespace.rs
