+++ crates/oak_sources/Cargo.toml
@@ -14,7 +14,7 @@ flate2.workspace = true
 hex.workspace = true
 log.workspace = true
 oak_fs.workspace = true
-oak_package.workspace = true
+oak_package_metadata.workspace = true
 oak_r_process.workspace = true
 serde.workspace = true
 serde_json.workspace = true
